Add files via upload
It is a report of Simulation-a-BPSK-modulated-baseband-system-with-Pulse-Shaping-Filter and analyzing the effect of roll-off factor using an eye diagram. The information has the Matlab code in it. You can refer to it. If you don't understand any part, feel free to comment or reach me at sanketzanwar1128@gmail.com, I will try my best to solve your doubts.